backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - Use psycopg2-binary for Windows compatibility
psycopg2-binary>=2.9.0

# SQLAlchemy and migrations - Updated for Python 3.13 compatibility
sqlalchemy>=2.0.25
alembic>=1.13.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling and utilities
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
python-slugify==8.0.1

# Image processing - Use latest compatible version
pillow>=9.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Data generation for sample data
faker>=20.0.0
